requirements.txt
asgiref==3.8.1
click==8.1.8
Django==5.2
django-cors-headers==4.7.0
djangorestframework==3.16.0
gensim==4.3.3
joblib==1.4.2
nltk==3.9.1
numpy==1.26.4
python-crfsuite==0.9.11
pyvi==0.1.1
regex==2024.11.6
scikit-learn==1.6.1
scipy==1.13.1
sklearn-crfsuite==0.5.0
smart-open==7.1.0
sqlparse==0.5.3
tabulate==0.9.0
threadpoolctl==3.6.0
tqdm==4.67.1
wrapt==1.17.2
